Updated to float types

+++ SharpQMapParser/Core/Plane.cs
@@ -6,8 +6,8 @@ namespace SharpQMapParser.Core
     {
         public Point[] Points = new Point[3];
         public string TextureName = string.Empty;
-        public int XOff, YOff;
-        public int Rotation;
+        public float XOff, YOff;
+        public float Rotation;
         public float XScale, YScale;
     }
 }
